ADC Voltage implimented

+++ FInalImplementation/Software/Project/Final_Project/Final_Project/uart.h
@@ -13,7 +13,7 @@ void usart_init(uint32_t baud_rate);
 void usart_transmit(char data);
 void usart_transmit_array(char* msg);
 void usart_breakdown_ascii(uint16_t number);
-void usart_transmit_voltage(uint16_t Voltage);
+void usart_transmit_voltage(int32_t Voltage);
 void usart_transmit_current(uint16_t Current);
 void usart_transmit_power(uint16_t power);
